+++ lib/rmega/crypto.rb
@@ -1,0 +1,13 @@
+require 'rmega/crypto/aes_ecb'
+require 'rmega/crypto/aes_cbc'
+require 'rmega/crypto/aes_ctr'
+require 'rmega/crypto/rsa'
+
+module Rmega
+  module Crypto
+    include AesCbc
+    include AesEcb
+    include AesCtr
+    include Rsa
+  end
+end