+++ lib/rmega/nodes/deletable.rb
@@ -1,6 +1,3 @@
-require 'rmega/utils'
-require 'rmega/crypto/crypto'
-
 module Rmega
   module Nodes
     module Deletable