+++ lib/rmega/nodes/folder.rb
@@ -1,10 +1,3 @@
-require 'rmega/crypto/crypto'
-require 'rmega/utils'
-require 'rmega/nodes/node'
-require 'rmega/nodes/expandable'
-require 'rmega/nodes/traversable'
-require 'rmega/nodes/deletable'
-
 module Rmega
   module Nodes
     class Folder < Node