+++ lib/rmega/nodes/root.rb
@@ -1,7 +1,3 @@
-require 'rmega/nodes/node'
-require 'rmega/nodes/expandable'
-require 'rmega/nodes/traversable'
-
 module Rmega
   module Nodes
     class Root < Node